Show latest revision in status message given multiple sources
In a setup with multiple sources the status message will only show an asterisk (`*`), in order to
help a user understand what version they are about to sync we do want to show the latest
available version. So we're adding a new condition an add this behaviour. For now we can assume
the revision we're looking for is always the revision of the first source.